// src/generated/models/JiraModel.ts
export interface FullIssue {
  // Unique id of the issue.
  id?: string;
  // Unique key of the issue.
  key?: string;
  // Browse to the issue using this URL.
  self?: string;
  fields?: {
    issuetype?: {
      // Unique id of the issue type.
      id?: string;
      // Verbose title of the issue type.
      description?: string;
      // Icon associated with the issue type.
      iconUrl?: string;
      // Title of the issue type.
      name?: string;
    };
    // The time spent on an issue
    timespent?: number;
    project?: {
      // The unique id of the project.
      id?: string;
      // The unique key of the project.
      key?: string;
      // Title of the project.
      name?: string;
      // Unique key of the project type.
      projectTypeKey?: string;
    };
    // The aggregate time spent on sub-tasks.
    aggregatetimespent?: number;
    // Type of resolution the issue has achieved.
    resolution?: {
      self?: string;
      id?: string;
      description?: string;
      name?: string;
    };
    // yyyy-MM-ddTHH:mm:ss.fffZ
    resolutiondate?: string;
    // The percentage of work logged vs the issue estimate.
    workratio?: number;
    // yyyy-MM-ddTHH:mm:ss.fffZ
    created?: string;
    priority?: {
      // Icon associated with the issue priority.
      iconUrl?: string;
      // Title of the priority.
      name?: string;
      // Id of the issue priority.
      id?: string;
    };
    // Time remaining estimated time in seconds.
    timeestimate?: number;
    // The original sum of all sub-task time estimates in seconds.
    aggregatetimeoriginalestimate?: number;
    assignee?: {
      // Person a issue is assigned to.
      accountId?: string;
      // User key of the person issue is assigned to.
      key?: string;
      // Email of the person issue is assigned to.
      emailAddress?: string;
      // Display name of the person issue is assigned to.
      displayName?: string;
    };
    // yyyy-MM-ddTHH:mm:ss.fffZ
    updated?: string;
    status?: {
      // Issue status.
      description?: string;
      // Issue status.
      iconUrl?: string;
      // Issue status.
      name?: string;
      // Issue status.
      id?: string;
      statusCategory?: {
        // Issue status category.
        id?: number;
        // Issue status category.
        key?: string;
        // Issue status category.
        colorName?: string;
        // Issue status category.
        name?: string;
      };
    };
    // The original time estimate in seconds.
    timeoriginalestimate?: number;
    // Issue description.
    description?: string;
    // Time sum of all sub-tasks remaining estimated time in seconds.
    aggregatetimeestimate?: number;
    // Title of the issue.
    summary?: string;
    // A system field that is multiple values addressed by 'name' (e.g. Microsoft Entra ID, Network Switch).
    components?: {
     id?: string;
     name?: string;
   }[];
    creator?: {
      // User who created the issue.
      accountId?: string;
      // Unique key of the user who created the issue.
      key?: string;
      // Email of the user who created the issue.
      emailAddress?: string;
      // Name of the user who created the issue.
      displayName?: string;
    };
    reporter?: {
      // User who reported the issue.
      accountId?: string;
      // Unique key of the user who reported the issue.
      key?: string;
      // Email of the user who reported the issue.
      emailAddress?: string;
      // Display name of the user who reported the issue.
      displayName?: string;
    };
    aggregateprogress?: {
      // The total progress completed of all sub-tasks in seconds.
      progress?: number;
      // The total sum of all estimated sub-task effort.
      total?: number;
      // The percent of aggregate completed progress in relation to estimated effort.
      percent?: number;
    };
    // yyyy-MM-ddTHH:mm:ss.fffZ
    duedate?: string;
    progress?: {
      // The progress complete in seconds.
      progress?: number;
      // The estimated effort.
      total?: number;
      // The percent of completed progress in relation to estimated effort.
      percent?: number;
    };
    // Epic name is required for epic issue type. This field matches 'customfield_10011' field on Jira server.
    customfield_10119?: unknown;
  };
}